+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import SearchBar from './components/SearchBar';
+import SearchBar from './components/SearchBar/SearchBar';
 
 function App() {
     return (
