+++ src/components/SearchBar/SearchInput.jsx
@@ -1,0 +1,12 @@
+const SearchInput = ({ value, onChange }) => {
+    return (
+        <input
+            type="text"
+            value={value}
+            onChange={onChange}
+            className="search-input"
+        />
+    );
+};
+
+export default SearchInput;
