+++ src/components/SearchBar/SearchItem.jsx
@@ -1,0 +1,11 @@
+import splitIntoChars from '../../utils/splitIntoChars';
+
+const SearchItem = ({ description, searchTerm }) => {
+    return (
+        <li className="group-item">
+            {splitIntoChars(description, searchTerm)}
+        </li>
+    );
+};
+
+export default SearchItem;
